test(alert): recategorize and clean up browser tests
